+++ src/FUMCard.js
@@ -15,7 +15,7 @@ import { Card, Table } from "react-bootstrap";
 import { decimalPlaces, stringMul } from "./utils";
 import { loadMetamask } from "./redux/interactions";
 import { buyFumBuilder, sellFumBuilder } from "./blockchainInteractions";
-import ecosystems from "./tokens";
+import ecosystems from "./ecosystems";
 import BlockchainWriteButtons from "./BlockchainWriteButtons";
 
 const FUMCard = ({
